requirements.txt
streamlit
pycountry
pymongo
python-dotenv
requests
schedule
